Okay, this patch REMOVES the -particles command line option, and adds a new cvar: cl_max_particles. This cvar is archived, has no lower or upper limits (well, less than 1 is not allowed) and can be changed in game at any time.
BUGS:

Only one so far. I can't figure out why it's doing this, but in software
clients, (well, at least X11) if you set it to 1 particle, it acts like
you set it far higher. 2 acts like you set it to zero... Or maybe
it's showing 2 and I just can't see it on my 320x200 window. In any
case, the vagary must be something in the software particles code,
because I basically used the same code from the GL particles code for
this as I used for the software renderer.

If nobody can find fault with my code, I'll just make a special note in
the console help.

In any case, let me know of any problems.

Misty-chan