Fixed a typo (cl_max_particles default is 2048!) as well as added a cvar I didn't know about (gl_lightmap_components)
Other than that, haven't added or removed much of anything - HOWEVER
note that this config has radically changed speedwise. I've gotten a
hardware upgrade, and so changed my config around to suit it. This is
written now for a voodoo 3 2000 pci, the old voodoo 1 has been
accidentally fried and given a proper burial.

If you relied on the old config to be fast as heck with it's defaults,
sorry. - Please look through the file, read the comments, and change the
settings to however you please.

Misty-chan