taniwha, please have a look at this to make sure I'm doing it the "right" way..  Sys_Printf is now Sys_StdPrintf for now.  Sys_Printf is a function pointer and will end up replacing Con_Printf as well.  Names will change to something intelligent when I'm done.  Win32 is not yet touched, but I haven't forgotten it.
The reason for this is the assumption that Con_Printf and Sys_Printf do
not overlap.  The new server console WILL change that, requiring that we
change how we handle both functions.  The plan:  Default to using stubs
and swap them for full-featured functions on Console init which do all
the things we expect.  Will also do a stderr version later.