This should fix bans and heartbeats to work with realtime resets.
Also of note, I found a line I missed for logging from the last checkin. Doh.

Did not test heartbeats, although the code is so dead simple it ought to
function.

I did not test logging, although the code is rather simple so unless I missed
something, it ought to work this time. ;)

Tim McGrath (Misty)