Proof of concept for making docs build without main configure
We need to figure out a way to make docs in unusual places, like in the
SourceForge shell environment -- which doesn't have an extremely recent
autotools. This is an extremely stripped-down automake/autoconf setup that
provides the bare minimum needed to AC_SUBST variables.

It's dirty and hacky, but it kinda works.